♻️ Refactor constructors to use new primary constructors syntax (#214)
* Refactor constructors to use new primary constructors syntax

Constructors for a several classes across the project have been refactored to take advantage of shortcut syntax. The changes reduce the amount of boilerplate code and increase readability. This also brings the code in line with modern conventions.

* Remove unnecessary logger declaration

The private logger declaration was removed from LoggingBehaviour.cs. It was unnecessary as the logger was being used directly, making the code more streamlined and lightweight.