🐛 148 fix file names of command item template (#222)
* Rename CommandNameValidator.cs to CommandNameCommandValidator.cs

In this commit, the CommandNameValidator.cs file has been renamed to CommandNameCommandValidator.cs. This is so that the file is correctly named when using `dotnet new ssw-ca-cmd`.

* Replace 'namespaceRoot' with 'slnName' in template configurations

In both command and query templates, the parameter 'namespaceRoot' has been replaced with 'slnName'. This new parameter is set as optional and it determines the solution name to be used for namespace prefix. The README file was updated to reflect these changes in the dotnet new commands.