Updated PowerLevelUpdatedEvent class in Heroes namespace
Refactored the `PowerLevelUpdatedEvent` class in the `SSW.CleanArchitecture.Domain.Heroes` namespace. Removed `OldPowerLevel` and `NewPowerLevel` properties and introduced a new `PowerDifference` property. The constructor has been updated to calculate and assign the `PowerDifference` instead of `OldPowerLevel` and `NewPowerLevel`. Also, the guards against null and negative values for `oldPowerLevel` and `newPowerLevel` have been removed from the constructor.